all the classes is setted and a first implementation of a list is working on.
