+++ app/build.gradle
@@ -23,10 +23,9 @@ dependencies {
     compile project(":volley")
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:support-v4:24.2.0'
-    // Important - the CameraSource implementation in this project requires version 8.1 or higher.
+// Important - the CameraSource implementation in this project requires version 8.1 or higher.
     compile 'com.google.android.gms:play-services-vision:9.4.0+'
     compile 'com.android.support:design:24.2.0'
 
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'}
 
-}
