+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.google.android.gms.samples.vision.barcodereader"
-    android:installLocation="auto" >
+    android:installLocation="auto">
 
     <uses-feature android:name="android.hardware.camera" />
+
     <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -13,8 +14,8 @@
         android:hardwareAccelerated="true"
         android:icon="@drawable/icon"
         android:label="MultiTrackerApp"
-        android:supportsRtl = "true"
-        android:theme="@style/Theme.AppCompat" >
+        android:supportsRtl="true"
+        android:theme="@style/Theme.AppCompat">
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
@@ -24,18 +25,23 @@
 
         <activity
             android:name=".VolleyActivity"
-            android:label="@string/title_activity_main" >
+            android:label="@string/title_activity_main">
+
+        </activity>
+        <activity
+            android:name=".BarcodeCaptureActivity"
+            android:label="Read Barcode" />
+        <activity
+            android:name=".MainActivity"
+            android:label="Read Barcode" />
+        <activity android:name=".ShowProductByCattegoryActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
 
-        <activity android:name=".BarcodeCaptureActivity"
-            android:label="Read Barcode"/>
-        <activity android:name=".MainActivity"
-            android:label="Read Barcode"/>
+        </activity>
     </application>
 
 </manifest>
