--- app/src/main/java/com/google/android/gms/samples/vision/barcodereader/Position.java
@@ -1,10 +1,0 @@
-package com.google.android.gms.samples.vision.barcodereader;
-
-/**
- * Created by giuliopettenuzzo on 16/05/17.
- */
-
-public interface Position {
-    public void setName(String name);
-    public String getName();
-}
