+++ app/src/main/java/com/google/android/gms/samples/vision/barcodereader/VolleyActivity.java
@@ -7,7 +7,6 @@ import android.graphics.BitmapFactory;
 import android.graphics.drawable.Drawable;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.os.Parcelable;
 import android.provider.MediaStore;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
@@ -22,7 +21,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.google.android.gms.vision.barcode.Barcode;
+import com.google.android.gms.samples.vision.barcodereader.urlManagement.ImageUrlUnpacker;
 
 import java.io.InputStream;
 
@@ -168,10 +167,7 @@ public class VolleyActivity extends AppCompatActivity {
             }
         });
         loadAllImage();
-        Product product = new RealProduct("ciao","ciao");
-        Intent inten = new Intent();
-        inten.putExtra("ciao", product);
-        Parcelable x = product;
+
 
     }
