+++ app/src/main/java/com/google/android/gms/samples/vision/barcodereader/entities/Attribute.java
@@ -1,0 +1,12 @@
+package com.google.android.gms.samples.vision.barcodereader.entities;
+
+/**
+ * Created by giuliopettenuzzo on 16/05/17.
+ */
+
+public interface Attribute {
+    public String getName();
+    public void setName(String name);
+    public String getValue();
+    public void setValue(String value);
+}
