+++ app/src/main/java/com/google/android/gms/samples/vision/barcodereader/entities/Position.java
@@ -1,0 +1,10 @@
+package com.google.android.gms.samples.vision.barcodereader.entities;
+
+/**
+ * Created by giuliopettenuzzo on 16/05/17.
+ */
+
+public interface Position {
+    public void setName(String name);
+    public String getName();
+}
