+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,5 @@
     <string name="barcode_success">Barcode read successfully</string>
     <string name="barcode_failure">No barcode captured</string>
     <string name="barcode_error">"Error reading barcode: %1$s"</string>
+    <string name="title_activity_show_product_by_cattegory">ShowProductByCattegory</string>
 </resources>
